refactor: Consolidate edge worker into API and improve deployment architecture
