+++ .env
@@ -23,7 +23,7 @@ DATABASE_URL=postgres://postgres:postgres@localhost:5432/example
 
 # Cloudflare Hyperdrive for local development
 # https://developers.cloudflare.com/hyperdrive/
-WRANGLER_HYPERDRIVE_LOCAL_CONNECTION_STRING_HYPERDRIVE=postgres://postgres:postgres@localhost:5432/example
+WRANGLER_HYPERDRIVE_LOCAL_CONNECTION_STRING_HYPERDRIVE_CACHED=postgres://postgres:postgres@localhost:5432/example
 WRANGLER_HYPERDRIVE_LOCAL_CONNECTION_STRING_HYPERDRIVE_DIRECT=postgres://postgres:postgres@localhost:5432/example
 
 # Better Auth
