+++ .gemini/settings.json
@@ -1,0 +1,4 @@
+{
+  "preferredEditor": "vscode",
+  "contextFileName": "CLAUDE.md"
+}
