+++ apps/app/package.json
@@ -8,7 +8,9 @@
     "build": "vite build",
     "preview": "vite preview",
     "test": "vitest",
-    "coverage": "vitest --coverage"
+    "coverage": "vitest --coverage",
+    "deploy": "wrangler deploy --env-file ../../.env --env-file ../../.env.local",
+    "logs": "wrangler tail --env-file ../../.env --env-file ../../.env.local"
   },
   "dependencies": {
     "@radix-ui/react-checkbox": "^1.3.3",
@@ -19,7 +21,7 @@
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-switch": "^1.2.6",
     "@repo/ui": "workspace:*",
-    "@tanstack/react-router": "^1.131.27",
+    "@tanstack/react-router": "^1.131.28",
     "better-auth": "^1.3.7",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -34,13 +36,13 @@
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
     "@tailwindcss/postcss": "^4.1.12",
-    "@tanstack/router-devtools": "^1.131.27",
-    "@tanstack/router-plugin": "^1.131.27",
+    "@tanstack/router-devtools": "^1.131.28",
+    "@tanstack/router-plugin": "^1.131.28",
     "@types/bun": "^1.2.21",
     "@types/node": "^24.3.0",
-    "@types/react": "^19.1.11",
-    "@types/react-dom": "^19.1.8",
-    "@vitejs/plugin-react": "^5.0.1",
+    "@types/react": "^19.1.12",
+    "@types/react-dom": "^19.1.9",
+    "@vitejs/plugin-react": "^5.0.2",
     "@vitejs/plugin-react-swc": "^4.0.1",
     "autoprefixer": "^10.4.21",
     "envars": "^1.1.1",
@@ -53,6 +55,7 @@
     "typescript": "~5.9.2",
     "vite": "~7.1.3",
     "vite-tsconfig-paths": "^5.1.4",
-    "vitest": "~3.2.4"
+    "vitest": "~3.2.4",
+    "wrangler": "^4.33.1"
   }
 }
