--- apps/edge/README.md
@@ -1,1 +1,0 @@
-# Edge Server
