--- apps/edge/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@repo/edge",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "bun run build && bun --cwd ../.. wrangler dev --config apps/edge/wrangler.jsonc",
-    "dev:watch": "vite build --watch & bun --cwd ../.. wrangler dev --config apps/edge/wrangler.jsonc",
-    "build": "vite build",
-    "test": "vitest",
-    "coverage": "vitest run --coverage",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@hono/zod-validator": "^0.7.2",
-    "@repo/api": "workspace:*",
-    "@repo/core": "workspace:*",
-    "@repo/db": "workspace:*",
-    "@trpc/server": "^11.5.0",
-    "better-auth": "^1.3.7",
-    "hono": "^4.9.4",
-    "zod": "^4.1.3"
-  },
-  "peerDependencies": {
-    "@cloudflare/workers-types": "^4.20250826.0"
-  },
-  "peerDependenciesMeta": {
-    "@cloudflare/workers-types": {
-      "optional": true
-    }
-  },
-  "devDependencies": {
-    "@cloudflare/vite-plugin": "^1.11.7",
-    "@cloudflare/vitest-pool-workers": "^0.8.66",
-    "@cloudflare/workers-types": "^4.20250826.0",
-    "@hono/vite-build": "^1.7.0",
-    "@repo/typescript-config": "workspace:*",
-    "@types/bun": "^1.2.21",
-    "typescript": "~5.9.2",
-    "vite": "~7.1.3",
-    "vitest": "~3.2.4"
-  }
-}
