--- apps/edge/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../packages/typescript-config/cloudflare.jsonc",
-  "compilerOptions": {
-    "composite": true,
-    "declaration": true,
-    "noEmit": false,
-    "outDir": "./dist",
-    "types": ["@cloudflare/workers-types", "vite/client"]
-  },
-  "include": ["**/*.ts", "**/*.json", "**/*.d.ts"],
-  "exclude": ["dist", "node_modules"],
-  "references": [
-    { "path": "../api" },
-    { "path": "../../packages/core" },
-    { "path": "../../db" }
-  ]
-}
