--- apps/edge/types/cloudflare-env-overrides.d.ts
@@ -1,15 +1,0 @@
-/* SPDX-FileCopyrightText: 2014-present Kriasoft */
-/* SPDX-License-Identifier: MIT */
-
-declare namespace Cloudflare {
-  interface Env {
-    // Override generated union types to be more flexible for runtime processing
-    ALLOWED_ORIGINS: string;
-
-    // Secret bindings are not generated by `wrangler types`
-    BETTER_AUTH_SECRET: string;
-    GOOGLE_CLIENT_ID: string;
-    GOOGLE_CLIENT_SECRET: string;
-    OPENAI_API_KEY: string;
-  }
-}
