--- apps/edge/types/cloudflare-env.d.ts
@@ -1,27 +1,0 @@
-/* eslint-disable */
-// Generated by Wrangler by running `wrangler types --config apps/edge/wrangler.jsonc --include-runtime=false apps/edge/types/cloudflare-env.d.ts` (hash: 80851b1795c054c12e253fa47ed87bd0)
-declare namespace Cloudflare {
-  interface Env {
-    ENVIRONMENT: "development" | "production" | "staging" | "preview";
-    ALLOWED_ORIGINS:
-      | "http://localhost:5173,http://127.0.0.1:5173"
-      | "https://example.com"
-      | "https://staging.example.com"
-      | "https://preview.example.com";
-    HYPERDRIVE: Hyperdrive;
-    HYPERDRIVE_DIRECT: Hyperdrive;
-    ASSETS: Fetcher;
-  }
-}
-interface Env extends Cloudflare.Env {}
-type StringifyValues<EnvType extends Record<string, unknown>> = {
-  [Binding in keyof EnvType]: EnvType[Binding] extends string
-    ? EnvType[Binding]
-    : string;
-};
-declare namespace NodeJS {
-  interface ProcessEnv
-    extends StringifyValues<
-      Pick<Cloudflare.Env, "ENVIRONMENT" | "ALLOWED_ORIGINS">
-    > {}
-}
