--- apps/edge/types/index.ts
@@ -1,7 +1,0 @@
-/* SPDX-FileCopyrightText: 2014-present Kriasoft */
-/* SPDX-License-Identifier: MIT */
-
-import "./cloudflare-env";
-import "./cloudflare-env-overrides";
-
-export type CloudflareEnv = Cloudflare.Env;
