--- apps/edge/wrangler.jsonc
@@ -1,107 +1,0 @@
-{
-  "$schema": "node_modules/wrangler/config-schema.json",
-
-  "name": "example",
-  "main": "./dist/index.js",
-
-  "assets": {
-    "directory": "../../apps/web/dist",
-    "binding": "ASSETS"
-  },
-
-  "upload_source_maps": true,
-
-  "compatibility_date": "2025-05-05",
-  "compatibility_flags": ["nodejs_compat"],
-
-  // Default configuration applied to all environments unless overridden
-  "vars": {
-    "ENVIRONMENT": "development",
-    "ALLOWED_ORIGINS": "http://localhost:5173,http://127.0.0.1:5173"
-  },
-  "hyperdrive": [
-    {
-      "binding": "HYPERDRIVE",
-      "id": "your-hyperdrive-cached-id-here"
-    },
-    {
-      "binding": "HYPERDRIVE_DIRECT",
-      "id": "your-hyperdrive-direct-id-here"
-    }
-  ],
-  "kv_namespaces": [],
-
-  // Overrides and settings for each deployment environment
-  // (production, staging, preview)
-  "env": {
-    "production": {
-      "routes": [
-        {
-          "pattern": "example.com",
-          "custom_domain": true
-        }
-      ],
-      "vars": {
-        "ENVIRONMENT": "production",
-        "ALLOWED_ORIGINS": "https://example.com"
-      },
-      "hyperdrive": [
-        {
-          "binding": "HYPERDRIVE",
-          "id": "your-production-hyperdrive-cached-id-here"
-        },
-        {
-          "binding": "HYPERDRIVE_DIRECT",
-          "id": "your-production-hyperdrive-direct-id-here"
-        }
-      ],
-      "kv_namespaces": []
-    },
-    "staging": {
-      "routes": [
-        {
-          "pattern": "staging.example.com",
-          "custom_domain": true
-        }
-      ],
-      "vars": {
-        "ENVIRONMENT": "staging",
-        "ALLOWED_ORIGINS": "https://staging.example.com"
-      },
-      "hyperdrive": [
-        {
-          "binding": "HYPERDRIVE",
-          "id": "your-staging-hyperdrive-cached-id-here"
-        },
-        {
-          "binding": "HYPERDRIVE_DIRECT",
-          "id": "your-staging-hyperdrive-direct-id-here"
-        }
-      ],
-      "kv_namespaces": []
-    },
-    "preview": {
-      "routes": [
-        {
-          "pattern": "preview.example.com",
-          "custom_domain": true
-        }
-      ],
-      "vars": {
-        "ENVIRONMENT": "preview",
-        "ALLOWED_ORIGINS": "https://preview.example.com"
-      },
-      "hyperdrive": [
-        {
-          "binding": "HYPERDRIVE",
-          "id": "your-preview-hyperdrive-cached-id-here"
-        },
-        {
-          "binding": "HYPERDRIVE_DIRECT",
-          "id": "your-preview-hyperdrive-direct-id-here"
-        }
-      ],
-      "kv_namespaces": []
-    }
-  }
-}
