+++ package.json
@@ -19,22 +19,26 @@
     "test": "vitest",
     "typecheck": "tsc --build",
     "build:types": "tsc --build",
-    "build": "bun --filter @repo/web --filter @repo/api --filter @repo/edge build",
-    "build:api": "docker build --tag api:latest -f ./apps/api/Dockerfile .",
-    "generate:types": "bun wrangler types --config apps/edge/wrangler.jsonc --include-runtime=false apps/edge/types/cloudflare-env.d.ts && bun prettier --write apps/edge/types/*-env.d.ts",
+    "build": "bun --filter @repo/web --filter @repo/api build",
+    "generate:types": "bun wrangler types --config apps/api/wrangler.jsonc --include-runtime=false apps/api/types/cloudflare-env.d.ts && bun prettier --write apps/api/types/*-env.d.ts",
     "prepare": "husky",
     "docs:dev": "vitepress dev docs",
     "docs:build": "vitepress build docs",
     "docs:preview": "vitepress preview docs",
     "docs:deploy": "gh-pages --dist docs/.vitepress/dist",
-    "dev:web": "bun --filter @repo/web dev",
-    "dev:api": "bun --filter @repo/api dev",
-    "dev:edge": "bun --filter @repo/edge dev",
-    "build:web": "bun --filter @repo/web build",
-    "build:edge": "bun --filter @repo/edge build",
-    "test:web": "bun --filter @repo/web test",
-    "test:api": "bun --filter @repo/api test",
-    "test:edge": "bun --filter @repo/edge test",
+    "app:dev": "bun --filter @repo/app dev",
+    "app:build": "bun --filter @repo/app build",
+    "app:test": "bun --filter @repo/app test",
+    "app:deploy": "bun --filter @repo/app deploy",
+    "web:dev": "bun --filter @repo/web dev",
+    "web:build": "bun --filter @repo/web build",
+    "web:test": "bun --filter @repo/web test",
+    "web:deploy": "bun --filter @repo/web deploy",
+    "api:dev": "bun --filter @repo/api dev",
+    "api:build": "bun --filter @repo/api build",
+    "api:build:docker": "docker build --tag api:latest -f ./apps/api/Dockerfile .",
+    "api:test": "bun --filter @repo/api test",
+    "api:deploy": "bun --filter @repo/api deploy",
     "ui:add": "bun --filter @repo/ui ui:add",
     "ui:list": "bun --filter @repo/ui ui:list",
     "ui:update": "bun --filter @repo/ui ui:update",
@@ -43,7 +47,7 @@
   "devDependencies": {
     "@emotion/babel-plugin": "^11.13.5",
     "@emotion/eslint-plugin": "^11.12.0",
-    "@eslint-react/eslint-plugin": "^1.52.6",
+    "@eslint-react/eslint-plugin": "^1.52.8",
     "@eslint/js": "^9.34.0",
     "@types/eslint": "^9.6.1",
     "@typescript-eslint/eslint-plugin": "^8.41.0",
@@ -73,7 +77,7 @@
     "vitepress": "^1.6.4",
     "vitepress-plugin-mermaid": "^2.0.17",
     "vitest": "~3.2.4",
-    "wrangler": "^4.32.0"
+    "wrangler": "^4.33.1"
   },
   "prettier": {
     "printWidth": 80,
