+++ packages/ui/package.json
@@ -48,8 +48,8 @@
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
-    "@types/react": "^19.1.11",
-    "@types/react-dom": "^19.1.8",
+    "@types/react": "^19.1.12",
+    "@types/react-dom": "^19.1.9",
     "execa": "^9.6.0",
     "globby": "^14.1.0",
     "tailwindcss": "^4.1.12",
