+++ packages/ws-protocol/package.json
@@ -15,13 +15,13 @@
   },
   "peerDependencies": {
     "bun-ws-router": "^0.3.2",
-    "zod": "^4.1.3"
+    "zod": "^4.1.5"
   },
   "devDependencies": {
     "@repo/typescript-config": "workspace:*",
     "@types/bun": "^1.2.21",
     "bun-ws-router": "^0.3.2",
     "typescript": "~5.9.2",
-    "zod": "^4.1.3"
+    "zod": "^4.1.5"
   }
 }
