+++ tsconfig.json
@@ -2,7 +2,7 @@
   "files": [],
   "references": [
     { "path": "./apps/api" },
-    { "path": "./apps/edge" },
+    { "path": "./apps/app" },
     { "path": "./apps/web" },
     { "path": "./db" },
     { "path": "./packages/core" },
